Move variable-specific stuff into variables module
