+++ packages/shadenfreude/src/index.ts
@@ -3,3 +3,4 @@ export * from "./factories"
 export * from "./helpers"
 export * from "./nodes"
 export * from "./types"
+export * from "./variables"
