+++ packages/shadenfreude/src/nodes/inputs/VertexPositionNode.ts
@@ -1,4 +1,4 @@
-import { nodeFactory } from "../../factories"
+import { nodeFactory } from "../.."
 import { Vec3VaryingNode } from "../util/Vec3VaryingNode"
 
 export const VertexPositionNode = nodeFactory(() => ({
