+++ packages/shadenfreude/src/nodes/math/DivideNode.ts
@@ -1,5 +1,4 @@
-import { nodeFactory } from "../../factories"
-import { Value } from "../../types"
+import { nodeFactory, Value } from "../.."
 import { OperatorNode } from "./OperatorNode"
 
 export const DivideNode = nodeFactory<{
