+++ packages/shadenfreude/src/nodes/values/Vector3Node.ts
@@ -1,5 +1,4 @@
-import { nodeFactory, vec3 } from "../../factories"
-import { Value } from "../../types"
+import { nodeFactory, Value, vec3 } from "../.."
 
 export const Vector3Node = nodeFactory<{ value: Value<"vec3"> }>(
   ({ value }) => ({
